refactor: extract auth deps from api/deps.py into auth/deps.py
Move authentication-related FastAPI dependencies to the new auth module:
  - bearer_scheme       — HTTPBearer token extraction
  - get_current_user()  — JWT verify → get_or_create_user
  - require_role()      — system role check factory

auth/deps.py imports from auth/cognito_token.py (new home) instead
of core/cognito.py.

api/deps.py now re-exports the auth deps for backward compatibility
and retains the org-scoped deps (get_current_org_membership,
require_org_role) which will move to orgs/deps.py in a later PR.
