+++ app/api/deps.py
@@ -1,3 +1,12 @@
+"""
+API dependencies — auth deps re-exported from auth.deps for backward compat.
+
+Auth deps (get_current_user, require_role, bearer_scheme) have moved to
+app.auth.deps. This module re-exports them so existing callers still work.
+Org deps (get_current_org_membership, require_org_role) remain here
+until they are migrated to app.orgs.deps in a later PR.
+"""
+
 from __future__ import annotations
 
 import uuid
@@ -5,92 +14,16 @@ from collections.abc import Callable
 from typing import Any
 
 from fastapi import Depends, Header, HTTPException, status
-from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.core.cognito import get_email_from_access_token, verify_token
+from app.auth.deps import (  # noqa: F401
+    bearer_scheme,
+    get_current_user,
+    require_role,
+)
 from app.core.database import get_db
 from app.models.org_membership import OrgMembership
 from app.models.user import User
-from app.services.user_service import get_or_create_user
-
-bearer_scheme = HTTPBearer(auto_error=False)
-
-
-async def get_current_user(
-    credentials: HTTPAuthorizationCredentials | None = Depends(bearer_scheme),
-    db: AsyncSession = Depends(get_db),
-) -> User:
-    """
-    FastAPI dependency: validate JWT → get or create user in DB.
-
-    Usage:
-        @router.get("/something")
-        async def something(user: User = Depends(get_current_user)):
-            print(user.id, user.email, user.system_role)
-
-    Returns a User model instance (not raw claims).
-    Raises 401 if token is missing, expired, or invalid.
-    """
-    if credentials is None:
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="Missing authorization header",
-            headers={"WWW-Authenticate": "Bearer"},
-        )
-
-    token = credentials.credentials
-
-    try:
-        claims: dict[str, Any] = verify_token(token)
-    except Exception:
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="Invalid or expired token",
-            headers={"WWW-Authenticate": "Bearer"},
-        )
-
-    # Cognito access tokens don't include the email claim (only ID tokens do).
-    # Fall back to a get_user API call so first-login user creation works.
-    email = claims.get("email", "") or get_email_from_access_token(token)
-
-    user = await get_or_create_user(
-        db=db,
-        cognito_sub=claims["sub"],
-        email=email,
-    )
-
-    return user
-
-
-def require_role(*allowed_roles: str) -> Callable[..., Any]:
-    """
-    Dependency factory that checks the user's system_role.
-
-    Usage:
-        from app.core.roles import SystemRole
-
-        @router.get("/admin-only")
-        async def admin_only(
-            user: User = Depends(require_role(SystemRole.ADMIN)),
-        ):
-            ...
-
-    Returns the User if they have one of the allowed roles.
-    Raises 403 if they don't.
-    """
-
-    async def _check_role(
-        current_user: User = Depends(get_current_user),
-    ) -> User:
-        if current_user.system_role not in allowed_roles:
-            raise HTTPException(
-                status_code=status.HTTP_403_FORBIDDEN,
-                detail="Insufficient permissions",
-            )
-        return current_user
-
-    return _check_role
 
 
 async def get_current_org_membership(
